src/requirements.txt
APScheduler==2.1.2
Flask==1.1.2
Flask_SQLAlchemy==2.2
PyJWT==1.7.1
